Refactor: BaseMviViewModel 적용
