+++ feature/pdf/src/main/java/kr/co/pdf/PdfViewModel.kt
@@ -2,8 +2,8 @@ package kr.co.pdf
 
 import android.graphics.pdf.PdfRenderer
 import kotlinx.coroutines.flow.collectLatest
-import kr.co.pdf.model.PdfUiIntent
-import kr.co.pdf.model.PdfUiState
+import kr.co.model.PdfUiIntent
+import kr.co.model.PdfUiState
 import kr.co.ui.base.BaseMviViewModel
 import kr.co.ui.util.TopBarState
 import kr.co.util.PdfToBitmap
