--- feature/pdf/src/main/java/kr/co/pdf/model/PdfUiIntent.kt
@@ -1,16 +1,0 @@
-package kr.co.pdf.model
-
-import android.graphics.pdf.PdfRenderer
-import kr.co.ui.base.UiIntent
-import kr.co.ui.util.TopBarState
-
-internal sealed class PdfUiIntent: UiIntent {
-    data class Init(
-        val renderer: PdfRenderer,
-        val topBarState: TopBarState
-    ) : PdfUiIntent()
-
-    data object ShowTopBar : PdfUiIntent()
-    data class ChangePage(val page: Int) : PdfUiIntent()
-    data class RenderPage(val page: Int) : PdfUiIntent()
-}
