+++ feature/recent/src/main/java/kr/co/model/RecentUiIntent.kt
@@ -1,0 +1,8 @@
+package kr.co.model
+
+import kr.co.ui.base.UiIntent
+
+internal sealed interface RecentUiIntent: UiIntent {
+    data object Init : RecentUiIntent
+    data class ClickFile(val file: FileInfo) : RecentUiIntent
+}
