+++ feature/recent/src/main/java/kr/co/recent/RecentViewModel.kt
@@ -1,28 +1,28 @@
 package kr.co.recent
 
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.collectLatest
-import kotlinx.coroutines.flow.update
-import kotlinx.coroutines.launch
 import kr.co.data.repository.RecentRepository
+import kr.co.model.RecentUiIntent
 import kr.co.model.RecentUiState
+import kr.co.ui.base.BaseMviViewModel
 
 internal class RecentViewModel(
     recentRepository : RecentRepository,
-) : ViewModel() {
+) : BaseMviViewModel<RecentUiState, RecentUiIntent>(RecentUiState.INIT) {
 
-    private val _uiState = MutableStateFlow(RecentUiState.EMPTY)
-    val uiState = _uiState.asStateFlow()
+    override fun handleIntent(intent: RecentUiIntent) {
+        when (intent) {
+            is RecentUiIntent.Init -> {}
+            is RecentUiIntent.ClickFile -> {}
+        }
+    }
 
     init {
-        viewModelScope.launch {
+        launch {
             recentRepository.get()
                 .collectLatest { files ->
-                    _uiState.update {
-                        it.copy(
+                    reduce {
+                        copy(
                             files = files
                         )
                     }
